% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNsamp.R
\name{simNsamp}
\alias{simNsamp}
\title{simNsamp: Simple Simulation and Sampling}
\usage{
simNsamp(groups = c("jack", "2ocean", "3ocean", "4ocean"),
  groupProps = c(0.1, 0.35, 0.54, 0.01), sampleSizes = seq(20, 100, 20),
  bootstraps = 100, simulations = 10, alpha = 0.1)
}
\arguments{
\item{groups}{the groups}

\item{groupProps}{the group proportions}

\item{sampleSizes}{the sample sizes to be evaluated}

\item{bootstraps}{the desired number of bootstraps}

\item{alpha}{the desired alpha}

\item{simulation}{the desired number of simulations}
}
\description{
A simple function to simulate a population and then sample that population at varying sample sizes to
determine sample sizes needed to achieve a given coefficient of variation
}
\examples{
simNsamp

}
\author{
Mike Ackerman
}

